Fix Int64 to Int32 casting errors - use Convert.ToInt32 for all SQL aggregates
- Changed ErrorCode from int to long (event_message_id is BIGINT)
- Updated GetErrorsAsync to use GetInt64 for error codes
- Added IsDBNull checks and Convert.ToInt32 for all COUNT/SUM operations
- Fixed metrics, trends, package performance, and failure patterns queries
- Prevents 'Unable to cast object of type System.Int64 to System.Int32' errors

+++ Models/ErrorLog.cs
@@ -5,7 +5,7 @@ namespace SSISAnalyticsDashboard.Models
         public long ExecutionId { get; set; }
         public string PackageName { get; set; } = string.Empty;
         public DateTime ErrorTime { get; set; }
-        public int ErrorCode { get; set; }
+        public long ErrorCode { get; set; }
         public string ErrorDescription { get; set; } = string.Empty;
     }
 }
